+++ src/main/java/safide/erp/inventory/domain/port/IInveProdCataRepository.java
@@ -3,5 +3,5 @@ package safide.erp.inventory.domain.port;
 import safide.erp.generic.domain.port.IGeneErpRepository;
 import safide.erp.inventory.domain.model.InveProdCata;
 
-public interface IInveProdCataRepository extends IGeneErpRepository<InveProdCata, Integer> {
+public interface IInveProdCataRepository extends IGeneErpRepository<InveProdCata, Long> {
 }
