+++ src/main/java/safide/erp/inventory/infrastructure/adapter/input/IInveProdCataCrudRepository.java
@@ -3,5 +3,5 @@ package safide.erp.inventory.infrastructure.adapter.input;
 import org.springframework.data.repository.CrudRepository;
 import safide.erp.inventory.infrastructure.entity.InveProdCataEntity;
 
-public interface IInveProdCataCrudRepository extends CrudRepository<InveProdCataEntity, Integer> {
+public interface IInveProdCataCrudRepository extends CrudRepository<InveProdCataEntity, Long> {
 }
