+++ src/main/java/safide/erp/inventory/infrastructure/entity/InveProdCataEntity.java
@@ -16,10 +16,10 @@ public class InveProdCataEntity {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "id_inprca")
-	private Integer id;
+	private Long id;
     @Column(unique = true, nullable = true)
 	private String name_inprca;
 	private String alterna_inprca;
-	private Integer id_macohe;
-	private Integer id_macode;
+	private Long id_macohe;
+	private Long id_macode;
 }
