+++ src/main/java/safide/erp/inventory/infrastructure/rest/InveProdHeadController.java
@@ -7,7 +7,7 @@ import safide.erp.inventory.domain.model.InveProdHead;
 
 @RestController
 @RequestMapping("/api/v1/products")
-public class InveProdHeadController extends GeneErpController<InveProdHead, Integer> {
+public class InveProdHeadController extends GeneErpController<InveProdHead, Long> {
     public InveProdHeadController(InveProdHeadService inveProdHeadService) {
         super(inveProdHeadService);
     }
