+++ src/main/java/safide/erp/management/domain/port/IManaCompDetaRepository.java
@@ -3,5 +3,5 @@ package safide.erp.management.domain.port;
 import safide.erp.generic.domain.port.IGeneErpRepository;
 import safide.erp.management.domain.model.ManaCompDeta;
 
-public interface IManaCompDetaRepository extends IGeneErpRepository<ManaCompDeta, Integer> {
+public interface IManaCompDetaRepository extends IGeneErpRepository<ManaCompDeta, Long> {
 }
