+++ src/main/java/safide/erp/management/infrastructure/adapter/input/IManaCompHeadCrudRepository.java
@@ -3,5 +3,5 @@ package safide.erp.management.infrastructure.adapter.input;
 import org.springframework.data.repository.CrudRepository;
 import safide.erp.management.infrastructure.entity.ManaCompHeadEntity;
 
-public interface IManaCompHeadCrudRepository extends CrudRepository<ManaCompHeadEntity, Integer> {
+public interface IManaCompHeadCrudRepository extends CrudRepository<ManaCompHeadEntity, Long> {
 }
