+++ src/main/java/safide/erp/sale/application/port/SaleCustHeadService.java
@@ -4,7 +4,7 @@ import safide.erp.generic.application.port.GeneErpService;
 import safide.erp.sale.domain.model.SaleCustHead;
 import safide.erp.sale.domain.port.ISaleCustHeadRepository;
 
-public class SaleCustHeadService extends GeneErpService<SaleCustHead, Integer> {
+public class SaleCustHeadService extends GeneErpService<SaleCustHead, Long> {
     public SaleCustHeadService(ISaleCustHeadRepository iSaleCustHeadRepository) {
         super(iSaleCustHeadRepository);
     }
