+++ src/main/java/safide/erp/sale/infrastructure/adapter/input/ISaleCustHeadCrudRepository.java
@@ -3,5 +3,5 @@ package safide.erp.sale.infrastructure.adapter.input;
 import org.springframework.data.repository.CrudRepository;
 import safide.erp.sale.infrastructure.entity.SaleCustHeadEntity;
 
-public interface ISaleCustHeadCrudRepository extends CrudRepository<SaleCustHeadEntity, Integer> {
+public interface ISaleCustHeadCrudRepository extends CrudRepository<SaleCustHeadEntity, Long> {
 }
